requirements.txt
google-genai
python-dotenv
rich
jinja2
pygments
